news_scraper.py
# news_scraper.py
from RPA.Browser.Selenium import Selenium
from datetime import datetime, timedelta
import openpyxl
import os
from config import OUTPUT_DIR, EXCEL_FILENAME

class NewsScraper:
    def __init__(self, search_phrase, news_category, num_months):
        self.search_phrase = search_phrase
        self.news_category = news_category
        self.num_months = num_months
        self.browser = Selenium()
        self.workbook = None
        self.worksheet = None

    def run(self):
        self.setup_excel()
        self.open_news_site()
        self.perform_search()
        self.select_category()
        self.scrape_news()
        self.save_excel()
        self.browser.close_all_browsers()

    def setup_excel(self):
        self.workbook = openpyxl.Workbook()
        self.worksheet = self.workbook.active
        self.worksheet.append(["Title", "Date", "Description", "Picture Filename", "Search Phrase Count", "Contains Money"])

    def open_news_site(self):
        self.browser.open_available_browser("https://www.example-news-site.com")

    def perform_search(self):
        # Implement search functionality

    def select_category(self):
        # Implement category selection

    def scrape_news(self):
        # Implement news scraping logic

    def save_excel(self):
        if not os.path.exists(OUTPUT_DIR):
            os.makedirs(OUTPUT_DIR)
        self.workbook.save(os.path.join(OUTPUT_DIR, EXCEL_FILENAME))